requirements.txt
discord.py
requests
python-binance
python-dotenv